ai-agent/requirements.txt
livekit-agents>=0.11.0
livekit-plugins-google>=0.8.0
livekit-plugins-silero>=0.7.0
python-dotenv>=1.0.1
httpx>=0.27.0
google-generativeai>=0.7.2
google-genai>=1.0.0
